# docs/api/swagger.yaml
basePath: /
definitions:
...
  github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.Code:
    enum:
    - INVALID_PAYLOAD
    - VALIDATION_FAILED
    - UNAUTHENTICATED
    - FORBIDDEN
    - RESOURCE_NOT_FOUND
    - CONFLICT
    - RATE_LIMITED
    - PAYLOAD_TOO_LARGE
    - SEARCH_UNAVAILABLE
    - INTERNAL_ERROR
    - NOT_IMPLEMENTED
    type: string
    x-enum-varnames:
    - CodeInvalidPayload
    - CodeValidationFailed
    - CodeUnauthenticated
    - CodeForbidden
    - CodeNotFound
    - CodeConflict
    - CodeRateLimited
    - CodePayloadTooLarge
    - CodeSearchUnavailable
    - CodeInternal
    - CodeNotImplemented
  github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.InvalidParam:
    properties:
      code:
        type: string
      field:
        type: string
    type: object
...
  github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails:
    properties:
      code:
        $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.Code'
      detail:
        type: string
      errors:
        items:
          $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.InvalidParam'
        type: array
      instance:
        type: string
      request_id:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
  github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.SuccessResponse:
    properties:
      data: {}
      request_id:
        type: string
    type: object
...
  internal_consent_delivery_http.grantConsentRequest:
    properties:
      document_type:
        type: string
      version:
        type: string
    type: object
  internal_consent_delivery_http.privacyRequestPayload:
    properties:
      request_type:
        type: string
    type: object
  internal_consent_delivery_http.unsubscribeRequest:
    properties:
      token:
        maxLength: 512
        type: string
    required:
    - token
    type: object
...
info:
  contact: {}
  description: Backend API for AquaWheel e-commerce project.
  title: AquaWheel Store API
  version: "1.0"
paths:
...
  /api/v1/consent/unsubscribe:
    post:
      consumes:
      - application/json
      description: For a guest contact with no account. The token is issued to one
        address, expires, and is the only thing that authorises the withdrawal.
      parameters:
      - description: Signed token from the email
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_consent_delivery_http.unsubscribeRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
      summary: Withdraw marketing consent using a signed link token
      tags:
      - Consent v1
...
  /api/v1/customers/me/consents:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
      summary: Read the authenticated customer's consent history
      tags:
      - Consent v1
    post:
      consumes:
      - application/json
      description: 'The version is part of the record: consent is to the text that
        was published, not to the document in general.'
      parameters:
      - description: Document type and version
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_consent_delivery_http.grantConsentRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
      summary: Record consent to a specific document version
      tags:
      - Consent v1
  /api/v1/customers/me/consents/{type}:
    delete:
      description: Withdrawing the terms a customer is served under is refused; the
        account has to be closed instead.
      parameters:
      - description: Document type
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
      summary: Withdraw a previously granted consent
      tags:
      - Consent v1
  /api/v1/customers/me/privacy-requests:
    post:
      consumes:
      - application/json
      description: Accepted for asynchronous handling; an administrator approves it
        before anything is exported or erased. Erasure returns 501 where the deployment
        has no erasure implementation configured.
      parameters:
      - description: Request type
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_consent_delivery_http.privacyRequestPayload'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
      summary: Submit a GDPR privacy request
      tags:
      - Consent v1
...
  /api/v1/legal/documents/active:
    get:
      description: 'Public: a visitor must be able to read the terms before consenting
        to them.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_VladHrytsaiuk_ecommerce-core_internal_http_apiresponse.ProblemDetails'
      summary: List the currently published legal documents
      tags:
      - Consent v1
...
securityDefinitions:
  bearerAuth:
    description: 'Введіть токен у форматі: Bearer {your_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...